CPCHEC-0001 Make Position an independent class

+++ FearlessJumper/app/src/main/java/com/stunapps/fearlessjumper/helper/EntityTransformCalculator.java
@@ -9,11 +9,11 @@ import com.stunapps.fearlessjumper.prefab.Prefab;
 
 public interface EntityTransformCalculator
 {
-    int getWidth(Entity entity);
+    float getWidth(Entity entity);
 
-    int getHeight(Entity entity);
+    float getHeight(Entity entity);
 
-    int getWidth(Prefab prefab);
+    float getWidth(Prefab prefab);
 
-    int getHeight(Prefab prefab);
+    float getHeight(Prefab prefab);
 }
